Fairly big changes to the mode system.
We want to be able to use the toolkit's stats system to determine
which translation units should be editable in the various modes
that are envisaged for VirTaal. For example, during Quick Translate
mode, the user should be able to quickly move from one untranslated
or fuzzy translation unit to the next. We use the toolkit's translation
stats system to tell us which of the units fall into this category.

The toolkit's stat system returns arrays of indices into a translation
unit store. In fact, it returns a dictionary with arrays, where the
arrays are indexed (in the dictionary) according to the property of 
the units which they point to. For example, we might have 
("fuzzy", [1, 2, 3]) in the dictionary, which indicates that 
units at indices 1, 2 and 3 are fuzzy.

Sometimes we want to be able to look at composite properties; thus, we
are interested in two or more arrays at one time. For this reason, the
UnionSetEnumerator was implemented. It keeps the union of various
SortedSets and updates itself whenever one of its sub-SortedSets is
updated (via the GObject signal mechanism).

Each mode is currently an extension of the UnionSetEnumerator. The methods
next() and prev() should be used by the keyboard navigation system to
discover which translation unit to move to.

svn path=/src/trunk/virtaal/; revision=7285