Followed Friedel's suggestion that the editor widget only be created once for every translation unit.
Once created, (in unit_renderer.py:do_start_editing), the editor widget
is cached under the private variable __editor of its corresponding
translation unit.

The next time do_start_editing is called, we simply re-use the widget
in __editor and just reset the heights of its sub-widgets so that it
will display correctly.

In order to enumerate the sub-widgets to set their sizes, it was necessary
for each layout object to keep references to its children and so this
functionality was also added.

svn path=/src/trunk/virtaal/; revision=7325