Fixed bug #389.
Instead of attaching the CTRL-up and CTRL-down accelerators to the accelerator
group of the Virtaal object (which is the top-level object of the program),
an accelerator group is created for a UnitGrid. This group is then added to the
main window in VirTaal.

This group is deleted when the UnitGrid is destroyed. It seems necessary to invoke
the destroy method explicitly for any GTK object removed from a GTK container.

svn path=/src/trunk/virtaal/; revision=7377