Use libproxy if available
This should vastly improve our ability to work with proxies on Linux,
especially when proxy settings change while Virtaal is running. We
might even support PAC now! In theory this might work on other platforms,
but only Linux is tested, and Linux seems to be the main target for
libproxy. Our proxy support on Windows and OSX should be much better
simply because urllib provides us with more on those platforms.