Fix unnecessary layout changes in selectviews
We've forever had the problem of selected rows being layed out differently.
There are two issues:
(1) The selected rows are drawn using widgets, and the others we draw
using pango Layouts. By setting the desired width for the Label, the
selected rows appear the same, except in the case of RTL :-(.
(2) The padding was not the same, causing slight horisontal shifting.

This fixes github issue 1067.