Collect all signals for parent GObjects
This fixes several TM backends. RESTRequest was not seen as having the
signal 'http-success' for example. This is inherrited from the parent
class HTTPRequest, but is not returned by GObject.signal_list_names().
GObjectWrapper then blocks the emit() for such a signal. I'm not sure
where in the platform upgrades this broke, though.